server/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.1
blinker==1.5
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cssselect==1.2.0
dnspython==2.2.1
fake-useragent==1.1.1
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
parse==1.19.0
pyee==8.2.2
pymongo==4.3.3
pyppeteer==1.0.2
pyquery==1.4.3
python-dotenv==0.21.0
requests==2.28.1
requests-html==0.10.0
sentry-sdk==1.14.0
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.13
w3lib==2.1.1
websockets==10.4
Werkzeug==2.2.2
zipp==3.11.0
